move parselogline to separate file
TODO: remove opchain - opinfo already has a next
