+++ src/rsrc_mgr/resourcelog.c
@@ -64,8 +64,6 @@ GNU licenses can be found at http://www.gnu.org/licenses/.
 
 //   -------------   INTERNAL DEFINITIONS    -------------
 
-#define MAX_BUFFER 8192 // maximum character buffer to be used for parsing/printing log lines
-                        //    program will abort if limit is exceeded when reading or writing
 #define RECORD_LOG_PREFIX "RESOURCE-RECORD-LOGFILE\n" // prefix for a 'record'-log
                                                       //    - only op starts, no completions
 #define MODIFY_LOG_PREFIX "RESOURCE-MODIFY-LOGFILE\n" // prefix for a 'modify'-log
@@ -132,424 +130,6 @@ static void cleanuplog(RESOURCELOG rsrclog, char destroy) {
    else {
       pthread_mutex_unlock(&rsrclog->lock);
    }
-}
-
-/**
- * Parse a new operation (or sequence of them) from the given logfile
- * @param int logfile : Reference to the logfile to parse a line from
- * @param char* eof : Reference to a character to be populated with an exit flag value
- *                    1 if we hit EOF on the file on a line division
- *                    -1 if we hit EOF in the middle of a line
- *                    zero otherwise
- * @return opinfo* : Reference to a new set of operation info structs (caller must free)
- * NOTE -- Under most failure conditions, the logfile offset will be returned to its original value.
- *         This is not the case if parsing reaches EOF, in which case, offset will be left there.
- */
-opinfo* parselogline(int logfile, char* eof) {
-   char buffer[MAX_BUFFER] = {0};
-   char* tgtchar = buffer;
-   off_t origoff = lseek(logfile, 0, SEEK_CUR);
-   if (origoff < 0) {
-      LOG(LOG_ERR, "Failed to identify current logfile offset\n");
-      return NULL;
-   }
-   // read in an entire line
-   // NOTE -- Reading one char at a time isn't very efficient, but we don't expect parsing of
-   //         logfiles to be a significant performance factor.  This approach greatly simplifies
-   //         char buffer mgmt.
-   ssize_t readbytes;
-   while ((readbytes = read(logfile, tgtchar, 1)) == 1) {
-      // check for end of line
-      if (*tgtchar == '\n') { break; }
-      // check for excessive string length
-      if (tgtchar - buffer >= MAX_BUFFER - 1) {
-         LOG(LOG_ERR, "Parsed line exceeds memory limits\n");
-         lseek(logfile, origoff, SEEK_SET);
-         *eof = 0;
-         return NULL;
-      }
-      tgtchar++;
-   }
-
-   if (*tgtchar != '\n') {
-      if (readbytes == 0) {
-         if (tgtchar == buffer) {
-            LOG(LOG_INFO, "Hit EOF on logfile\n");
-            *eof = 1;
-         }
-         else {
-            LOG(LOG_ERR, "Hit mid-line EOF on logfile\n");
-            *eof = -1;
-         }
-
-         return NULL;
-      }
-
-      LOG(LOG_ERR, "Encountered error while reading from logfile\n");
-      lseek(logfile, origoff, SEEK_SET);
-      *eof = 0;
-      return NULL;
-   }
-
-   *eof = 0; // preemptively populate with zero
-
-   // allocate our operation node
-   opinfo* op = malloc(sizeof(*op));
-   op->extendedinfo = NULL;
-   op->start = 0;
-   op->count = 0;
-   op->errval = 0;
-   op->next = NULL;
-   op->ftag.ctag = NULL;
-   op->ftag.streamid = NULL;
-
-   // parse the op type and extended info
-   char* parseloc = buffer;
-   if (strncmp(buffer, "DEL-OBJ ", 8) == 0) {
-      LOG(LOG_INFO, "Parsing a MARFS_DELETE_OBJ operation\n");
-      op->type = MARFS_DELETE_OBJ_OP;
-      parseloc += 8;
-
-      // allocate delobj_info
-      delobj_info* extinfo = calloc(1, sizeof(*extinfo));
-
-      // parse in delobj_info
-      if (strncmp(parseloc, "{ ", 2)) {
-         LOG(LOG_ERR, "Missing '{ ' header for DEL-OBJ extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 2;
-
-      char* endptr = NULL;
-      unsigned long long parseval = strtoull(parseloc, &endptr, 10);
-      if (endptr == NULL || *endptr != ' ') {
-         LOG(LOG_ERR, "DEL-OBJ extended info has unexpected char in prev_active_index string: '%c'\n", *endptr);
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      extinfo->offset = (size_t)parseval;
-
-      parseloc = endptr + 1;
-
-      if (strncmp(parseloc, "} ", 2)) {
-         LOG(LOG_ERR, "Missing '} ' tail for DEL-OBJ extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 2;
-
-      // attach delobj_info
-      op->extendedinfo = extinfo;
-   }
-   else if (strncmp(buffer, "DEL-REF ", 8) == 0) {
-      LOG(LOG_INFO, "Parsing a MARFS_DELETE_REF operation\n");
-      op->type = MARFS_DELETE_REF_OP;
-      parseloc += 8;
-      // allocate delref_info
-      delref_info* extinfo = calloc(1, sizeof(*extinfo));
-
-      // parse in delref_info
-      if (strncmp(parseloc, "{ ", 2)) {
-         LOG(LOG_ERR, "Missing '{ ' header for DEL-REF extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 2;
-
-      char* endptr = NULL;
-      unsigned long long parseval = strtoull(parseloc, &endptr, 10);
-      if (endptr == NULL || *endptr != ' ') {
-         LOG(LOG_ERR, "DEL-REF extended info has unexpected char in prev_active_index string: '%c'\n", *endptr);
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      extinfo->prev_active_index = (size_t)parseval;
-
-      parseloc = endptr + 1;
-
-      if (strncmp(parseloc, "DZ ", 3) == 0) {
-         extinfo->delzero = 1;
-      }
-      else if (strncmp(parseloc, "-- ", 3) == 0) {
-         extinfo->delzero = 0;
-      }
-      else {
-         LOG(LOG_ERR, "Encountered unrecognized DEL-ZERO value in DEL-REF extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 3;
-
-      if (strncmp(parseloc, "EOS", 3) == 0) {
-         extinfo->eos = 1;
-      }
-      else if (strncmp(parseloc, "CNT", 3) == 0) {
-         extinfo->eos = 0;
-      }
-      else {
-         LOG(LOG_ERR, "Encountered unrecognized EOS value in DEL-REF extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 3;
-
-      if (strncmp(parseloc, " } ", 3)) {
-         LOG(LOG_ERR, "Missing ' } ' tail for DEL-REF extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 3;
-
-      // attach delref_info
-      op->extendedinfo = extinfo;
-   }
-   else if (strncmp(buffer, "REBUILD ", 8) == 0) {
-      LOG(LOG_INFO, "Parsing a MARFS_REBUILD operation\n");
-      op->type = MARFS_REBUILD_OP;
-      parseloc += 8;
-
-      // parse in rebuild_info
-      rebuild_info* extinfo = NULL;
-      if (strncmp(parseloc, "{ ", 2) == 0) {
-         // allocate rebuild_info
-         extinfo = calloc(1, sizeof(*extinfo));
-         parseloc += 2;
-
-         char* endptr = parseloc;
-         while (*endptr != ' ' && *endptr != '\n') { endptr++; }
-         if (*endptr != ' ') {
-            LOG(LOG_ERR, "Failed to parse markerpath from REBUILD extended info\n");
-            free(extinfo);
-            free(op);
-            lseek(logfile, origoff, SEEK_SET);
-            return NULL;
-         }
-
-         *endptr = '\0'; // temporarily truncate string
-
-         extinfo->markerpath = strdup(parseloc);
-
-         if (extinfo->markerpath == NULL) {
-            LOG(LOG_ERR, "Failed to duplicate markerpath from REBUILD extended info\n");
-            free(extinfo);
-            free(op);
-            lseek(logfile, origoff, SEEK_SET);
-            return NULL;
-         }
-
-         *endptr = ' ';
-
-         parseloc = endptr + 1;
-
-         if (*parseloc != '}' && *parseloc != '\0') {
-            // possibly parse rtag value
-            endptr = parseloc;
-            while (*endptr != '\0' && *endptr != '\n' && *endptr != ' ') { endptr++; }
-            if (*endptr != ' ') {
-               LOG(LOG_ERR, "Failed to identify end of rtag marker in REBUILD extended info string\n");
-               free(extinfo->markerpath);
-               free(extinfo);
-               free(op);
-               lseek(logfile, origoff, SEEK_SET);
-               return NULL;
-            }
-
-            extinfo->rtag = calloc(1, sizeof(RTAG));
-
-            *endptr = '\0'; // truncate string to make rtag parsing easier
-
-            if (rtag_initstr(extinfo->rtag, parseloc)) {
-               LOG(LOG_ERR, "Failed to parse rtag value of REBUILD extended info: \"%s\"\n", parseloc);
-               free(extinfo->rtag);
-               free(extinfo->markerpath);
-               free(extinfo);
-               free(op);
-               lseek(logfile, origoff, SEEK_SET);
-               return NULL;
-            }
-
-            *endptr = ' ';
-            parseloc = endptr + 1;
-         }
-
-         if (strncmp(parseloc, "} ", 2)) {
-            LOG(LOG_ERR, "Missing '} ' tail for REBUILD extended info\n");
-            if (extinfo->rtag) {
-               rtag_free(extinfo->rtag);
-               free(extinfo->rtag);
-            }
-            free(extinfo->markerpath);
-            free(extinfo);
-            free(op);
-            lseek(logfile, origoff, SEEK_SET);
-            return NULL;
-         }
-
-         parseloc += 2;
-      }
-      else {
-         LOG(LOG_INFO, "Rebuild op is lacking extended info\n");
-      }
-      // attach rebuild_info
-      op->extendedinfo = extinfo;
-   }
-   else if (strncmp(buffer, "REPACK ", 7) == 0) {
-      LOG(LOG_INFO, "Parsing a MARFS_REPACK operation\n");
-      op->type = MARFS_REPACK_OP;
-      parseloc += 7;
-
-      // allocate repack_info
-      repack_info* extinfo = calloc(1, sizeof(*extinfo));
-
-      // parse in repack_info
-      if (strncmp(parseloc, "{ ", 2)) {
-         LOG(LOG_ERR, "Missing '{ ' header for REPACK extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 2;
-
-      char* endptr = NULL;
-      unsigned long long parseval = strtoull(parseloc, &endptr, 10);
-      if (endptr == NULL || *endptr != ' ') {
-         LOG(LOG_ERR, "REPACK extended info has unexpected char in totalbytes string: '%c'\n", *endptr);
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      extinfo->totalbytes = (size_t)parseval;
-      parseloc = endptr;
-      if (strncmp(parseloc, " } ", 3)) {
-         LOG(LOG_ERR, "Missing ' } ' tail for REPACK extended info\n");
-         free(extinfo);
-         free(op);
-         lseek(logfile, origoff, SEEK_SET);
-         return NULL;
-      }
-
-      parseloc += 3;
-
-      // attach repack_info
-      op->extendedinfo = extinfo;
-   }
-   else {
-      LOG(LOG_ERR, "Unrecognized operation type value: \"%s\"\n", buffer);
-      free(op);
-      lseek(logfile, origoff, SEEK_SET);
-      return NULL;
-   }
-
-   // parse the start value
-   if (*parseloc == 'S') {
-      op->start = 1;
-   }
-   else if (*parseloc != 'E') {
-      LOG(LOG_ERR, "Unexpected START string value: '%c'\n", *parseloc);
-      goto error;
-   }
-
-   if (parseloc[1] != ' ') {
-      LOG(LOG_ERR, "Unexpected trailing character after START value: '%c'\n", *(parseloc + 1));
-      goto error;
-   }
-
-   parseloc += 2;
-
-   // parse the count value
-   char* endptr = NULL;
-   unsigned long long parseval = strtoull(parseloc, &endptr, 10);
-   if (endptr == NULL || *endptr != ' ') {
-      LOG(LOG_ERR, "Failed to parse COUNT value with unexpected char: '%c'\n", *endptr);
-      goto error;
-   }
-
-   op->count = (size_t)parseval;
-
-   parseloc = endptr + 1;
-
-   // parse the errno value
-   long sparseval = strtol(parseloc, &endptr, 10);
-   if (endptr == NULL || *endptr != ' ') {
-      LOG(LOG_ERR, "Failed to parse ERRNO value with unexpected char: '%c'\n", *endptr);
-      goto error;
-   }
-
-   op->errval = (int)sparseval;
-
-   parseloc = endptr + 1;
-
-   // parse the NEXT value
-   char nextval = 0;
-   if (*(tgtchar - 1) == '-') {
-      if (*(tgtchar - 2) != ' ') {
-         LOG(LOG_ERR, "Unexpected char preceeds NEXT flag: '%c'\n", *(tgtchar - 2));
-         goto error;
-      }
-
-      nextval = 1; // note that we need to append another op
-      tgtchar -= 2; // pull this back, so we'll trim off the NEXT value
-   }
-
-   // parse the FTAG value
-   *tgtchar = '\0'; // trim the string, to make FTAG parsing easy
-   if (ftag_initstr(&op->ftag, parseloc)) {
-      LOG(LOG_ERR, "Failed to parse FTAG value of log line\n");
-      goto error;
-   }
-
-   // finally, parse in any subsequent linked ops
-   if (nextval) {
-      // NOTE -- Recursive parsing isn't the most efficient approach.
-      //         Simple though, and, once again, we don't expect logfile parsing to be a
-      //         significant performance consideration.
-      LOG(LOG_INFO, "Recursively parsing subsequent operation\n");
-      op->next = parselogline(logfile, eof);
-      if (op->next == NULL) {
-         LOG(LOG_ERR, "Failed to parse linked operation\n");
-         resourcelog_freeopinfo(op);
-         if (*eof == 0) {
-             lseek(logfile, origoff, SEEK_SET);
-         }
-         return NULL;
-      }
-   }
-
-   return op;
-
-  error:
-   resourcelog_freeopinfo(op);
-   lseek(logfile, origoff, SEEK_SET);
-   return NULL;
 }
 
 /**
